[Feat] #98 완료된 미션 검증 추가

+++ src/main/java/com/umc9th/areumdap/domain/mission/service/MissionCommandService.java
@@ -118,6 +118,11 @@ public class MissionCommandService {
     @Transactional
     public void deleteMission(Long userId, Long missionId) {
         Mission mission = getOwnedMissionForUpdate(userId, missionId);
+
+        if (mission.getMissionStatus() != MissionStatus.COMPLETED) {
+            throw new GeneralException(ErrorStatus.MISSION_NOT_COMPLETED);
+        }
+
         missionRepository.delete(mission);
     }
